Dramatically speed up speakers endpoint for state-level queries
This is achieved by creating another materialized view that rolls the
filterable data up to the state level, so there are 4x fewer rows.

This also includes a fix to add in missing areas for which there are no
rows of speaker data but which fall within the bounding box.