Fix bug with PUMA layers not being removed properly
On very slow connections, the PUMA outlines requests can take a long
time to load. If they took even longer than the main request for
speakers and geometries, those requests might finish much later, causing
the PUMA outline layers to be added after the fact, even if they were no
longer appropriate to show (user had zoomed out very far).

This makes it so that those PUMA outline requests are cancelled and the
layers removed as soon as the main request is made.