Fix deployment script
Having separate docker-compose.prod.yml files was causing different
containers to be spun up and generally leading to weird behavior. This
simplifies the whole process by just checking out the code for the first
deployment and then updating only the image numbers to deploy new code.

This should work fine as long as the docker-compose file and the deploy
script don't change often -- which they shouldn't.